Enhanced fconsole function that...
1. Displays the value of objects at the time of logging not at the time
of display.

If you log a dom element for example, it logs a pointer to the dynamic
object in the document and when you look at the log you get the current
state and not the state at the time it was logged. This version clones
objects before displaying them so that you see the object as it was at
the time of logging.

2. The current function converts objects to strings - and you lose the
information therein.

This version clones the objects and displays them in a collapsed form.

Test with: `fconsole("test", "me", "please", {a:'a'}, [1,2], document);`