Fix and enhance Admin Elements list
1. Fix issue where group filter is not cleared when you change forms.

2. Combine sorting of the order and group columns so that:
a. If you sort by group, the elements are ordered ascending within the
group (regardless of ASC/DESC for the group itself) and drag / drop
reordering is enabled;
b. If you sort by ordering, then sorting is by group and order but
ASC/DESC is respected.

3. Drag and drop is now limited to the group that the element is in
(standard joomla functionality enabled by adding the group id as a
special attribute to the row div).