Fix fabrikmodalrepeat in repeat groups reusing ids
The subform created for fmr fields needs to use the fmr name / id as a
prefix including the repeat number in order to avoid the ids for the
fields inside the fmr using the same ids when the repeat group is
repeated.

For the sake of commonality I have reused the Joomla subform naming
convention.

Also standardised use of requirejs.